Better fix for ring start event race condition.
The previous fix didn't take the case of short ring tones into
consideration. In those cases, the ringer is no longer ringing at the
time the continued ring event appears, which made the ring not being
played again.
Fix that by keeping track of whether the ringer is currently is supposed
to be ringing and playing continued ring tones in that case.

Change-Id: Ia3c2293a67a086a0085059b9e045335fcdf42ea5